% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_values.R
\name{simulate_values}
\alias{simulate_values}
\title{Simulates values from a probability distribution}
\usage{
simulate_values(sample_size, distribution, ...)
}
\arguments{
\item{sample_size}{A integer value larger than zero}

\item{distribution}{A character value specifying the probability distribution}

\item{...}{parameters of the probability distribution}
}
\value{
A numeric vector
}
\description{
Produces a \code{sample_size} number of values Using the random generation
function for the given \code{distribution}. Only the Normal, Binomial
or Poisson distribution are currently accepted.
}
\details{
For the Normal distribution no parameters are necessary, this defaults to the
Standard Normal distribution. However, the Poisson and Binomial distribution do
require parameters and an error is thrown if these are not provided.
}
\examples{
simulate_values(10, 'Normal')
simulate_values(100, 'Binomial', size=10, prob=0.5)
}
\seealso{
\code{\link{rnorm}}, \code{\link{rpois}}, \code{\link{rbinom}}
}

