% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torbie.R
\docType{package}
\name{torbie}
\alias{torbie}
\alias{torbie-package}
\title{torbie: A package for simulating values from probability distributions}
\description{
The torbie package provides one function which allows you to simulate
values from probability distributions
}

