% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_values.R
\name{simulate_values}
\alias{simulate_values}
\title{Simulates values from a probability distribution}
\usage{
simulate_values(sample_size, distribution, ...)
}
\arguments{
\item{sample_size}{A integer value larger than zero}

\item{distribution}{A character value specifying the probability distribution}

\item{...}{parameters of the probability distribution}
}
\value{
A numeric vector
}
\description{
Description here
}
\examples{
simulate_values(10, 'norm')
simulate_values(100, 'binom')
}

